Association recruitercandidate

+++ src/services/candidate/deleteCandidateById.js
@@ -3,14 +3,19 @@ import { candidatesModel, socialMediaModel, candidatesTechnologiesModel } from '
 export const deleteCandidateById = async (id) => {
 
     try {
+        if (id === undefined || id === "") return undefined
+
         const candidate = await candidatesModel.findByPk(id)
+
+        const tech = await candidate.getTechnologies({ through: candidatesTechnologiesModel })
         
-        const tech = await candidate.getTechnologies({ through: candidatesTechnologiesModel})
-        await candidate.removeTechnologies(tech, { through: candidatesTechnologiesModel })
-        
+        if (tech != undefined) {
+            await candidate.removeTechnologies(tech, { through: candidatesTechnologiesModel })
+        }
+
         await socialMediaModel.destroy({ where: { candidate_id: id } })
 
-        return await candidatesModel.destroy({ where: { id}})
+        return await candidatesModel.destroy({ where: { id } })
     } catch (e) {
         console.log(e.message)
     }
